- Se cambia a ansible la instalacion del prodenv - Se añade upgrade de instalaciones docker - Fix en task que seteaba docker user a la variable de entorno - Añado una nota de que mysql_root_password tiene que ser igual a l a que se uso para instalar fpbx e issabel - Cambio RTPENGINE_HOST de kamaialio-local.cfg de fqdn a IP - Se quita tarea de instalacion de sngrep - Elimino paquete de odbc duplicado para AIO - Fix en template de odbcinst.ini para instalaciones docker

+++ deploy/ansible/deploy.sh
@@ -39,7 +39,7 @@ arg1=$1
 OSValidation(){
   if [ -z $PIP ]; then
     os=`awk -F= '/^NAME/{print $2}' /etc/os-release`
-    if [ "$os" == '"CentOS Linux"' ]; then
+    if [ "$os" == '"CentOS Linux"' ] || [ "$os" == '"Issabel PBX"' ] || [ "$os" == '"Sangoma Linux"' ]; then
       echo "Downloading and installing epel-release repository"
       yum install epel-release -y
       echo "Installing python2-pip"
@@ -84,10 +84,14 @@ TagCheck() {
     tag="database"
   elif [ "$arg1" == "--docker-build" ]; then
     tag="docker_build"
+    BUILD_IMAGES=true
   elif [ "$arg1" == "--docker-deploy" ]; then
     tag="docker_deploy"
   elif [ "$arg1" == "--integration-tests" ]; then
     tag="all,integration-tests"
+  elif [ "$arg1" == "--docker-no-build" ]; then
+    tag="docker_build"
+    BUILD_IMAGES=false
   fi
 }
 
@@ -192,7 +196,7 @@ AnsibleExec() {
     fi
     echo "Beginning the Omnileads installation with Ansible, this installation process can last between 30-40 minutes"
     echo ""
-    ${ANSIBLE}-playbook $verbose $TMP_ANSIBLE/omnileads.yml --extra-vars "iface=$INTERFACE build_dir=$TMP_OMINICONTACTO repo_location=$REPO_LOCATION docker_root=$USER_HOME" --tags "$tag" -i $TMP_ANSIBLE/inventory
+    ${ANSIBLE}-playbook $verbose $TMP_ANSIBLE/omnileads.yml --extra-vars "iface=$INTERFACE build_dir=$TMP_OMINICONTACTO repo_location=$REPO_LOCATION docker_root=$USER_HOME build_images=$BUILD_IMAGES" --tags "$tag" -i $TMP_ANSIBLE/inventory
     ResultadoAnsible=`echo $?`
     if [ $ResultadoAnsible == 0 ];then
       echo "
@@ -240,7 +244,7 @@ OSValidation
 for i in "$@"
 do
   case $i in
-    --upgrade|-u|--install|-i|--kamailio|-k|--asterisk|-a|--omniapp|-o|--omnivoip|--dialer|-di|--database|-da|--change-network|-cnet|--change-passwords|-cp|--docker-build|--docker-deploy|--integration-tests)
+    --upgrade|-u|--install|-i|--kamailio|-k|--asterisk|-a|--omniapp|-o|--omnivoip|--dialer|-di|--database|-da|--change-network|-cnet|--change-passwords|-cp|--docker-no-build|--docker-build|--docker-deploy|--integration-tests)
       TagCheck
       shift
     ;;
@@ -255,12 +259,13 @@ do
               -a --asterisk: execute asterisk related tasks
               -da --database: execute tasks related to database
               -di --dialer: execute tasks related to dialer (Wombat Dialer)
-              --docker-deploy: deploy Omnileads in docker containers using docker-compose. See /deploy/docker/README.md
-              --docker-build: build Omnileads images. See /deploy/docker/CONTRIBUTING.md
-              -i --install: make a fresh install of Omnileads
-              -k --kamailio: execute kamailio related tasks
-              -o --omniapp: execute omniapp related tasks
-              -u --upgrade: make an upgrade of Omnileads version
+              --docker-deploy: deploy Omnileads in docker containers using docker-compose.
+              --docker-build: build and push Omnileads images to a registry. 
+              --docker-no-build: execute build images steps without building and pushing the images.
+              -i --install: make a fresh install of Omnileads.
+              -k --kamailio: execute kamailio related tasks.
+              -o --omniapp: execute omniapp related tasks.
+              -u --upgrade: make an upgrade of Omnileads version.
               --iface --interface: set the iface when you want omnileads services listening (JUST USE THIS OPTION WHEN INSTALLATION IS SELFHOSTED)
             "
       shift
@@ -276,29 +281,33 @@ do
     ;;
   esac
 done
-./keytransfer.sh $INTERFACE
-ResultadoKeyTransfer=`echo $?`
-  if [ "$ResultadoKeyTransfer" == 1 ]; then
-    echo "It seems that you don't have generated keys in the server you are executing this script"
-    echo "Try with ssh-keygen or check the ssh port configured in server"
-    rm -rf /var/tmp/servers_installed
-    exit 1
-  elif [ "$ResultadoKeyTransfer" == 2 ]; then
-    echo "#######################################################################"
-    echo "# The option --interface must be used only in selfhosted installation #"
-    echo "#######################################################################"
-    exit 1
-  elif [ "$ResultadoKeyTransfer" == 3 ]; then
-    echo "#####################################"
-    echo "# Option --interface must be passed #"
-    echo "#####################################"
-    exit 1
-  elif [ "$ResultadoKeyTransfer" == 4 ]; then
-    echo "###############################################################"
-    echo "# It seems you typed a wrong interface in --interface option  #"
-    echo "###############################################################"
-    exit 1
-  fi
+if [ "$arg1" == "--docker-build" ] || [ "$arg1" == "--docker-no-build" ]; then
+  echo ""
+else
+  ./keytransfer.sh $INTERFACE
+  ResultadoKeyTransfer=`echo $?`
+    if [ "$ResultadoKeyTransfer" == 1 ]; then
+      echo "It seems that you don't have generated keys in the server you are executing this script"
+      echo "Try with ssh-keygen or check the ssh port configured in server"
+      rm -rf /var/tmp/servers_installed
+      exit 1
+    elif [ "$ResultadoKeyTransfer" == 2 ]; then
+      echo "#######################################################################"
+      echo "# The option --interface must be used only in selfhosted installation #"
+      echo "#######################################################################"
+      exit 1
+    elif [ "$ResultadoKeyTransfer" == 3 ]; then
+      echo "#####################################"
+      echo "# Option --interface must be passed #"
+      echo "#####################################"
+      exit 1
+    elif [ "$ResultadoKeyTransfer" == 4 ]; then
+      echo "###############################################################"
+      echo "# It seems you typed a wrong interface in --interface option  #"
+      echo "###############################################################"
+      exit 1
+    fi
+fi
 AnsibleInstall
 CodeCopy
 VersionGeneration
